+++ src/Inline/Renderer/StrongRenderer.php
@@ -5,7 +5,7 @@
  *
  * (c) Colin O'Dell <colinodell@gmail.com>
  *
- * Original code based on the CommonMark JS reference parser (http://bitly.com/commonmarkjs)
+ * Original code based on the CommonMark JS reference parser (http://bitly.com/commonmark-js)
  *  - (c) John MacFarlane
  *
  * For the full copyright and license information, please view the LICENSE
@@ -33,6 +33,6 @@ class StrongRenderer implements InlineRendererInterface
             throw new \InvalidArgumentException('Incompatible inline type: ' . get_class($inline));
         }
 
-        return new HtmlElement('strong', array(), $htmlRenderer->renderInlines($inline->getInlineContents()));
+        return new HtmlElement('strong', array(), $htmlRenderer->renderInlines($inline->getChildren()));
     }
 }